auth_service/requirements.txt
Flask==3.1.0
Flask-Login>=0.6.2
Flask-Session==0.8.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.37
PyMySQL>=1.0.2
requests>=2.31.0
python-dotenv==1.0.1
cryptography==44.0.0
redis>=4.1.0
flask_cors==5.0.0